refactor: reorganize tax ID and Brazilian masks utilities
