+++ src/index.ts
@@ -3,8 +3,8 @@ export * from './logger'
 export * from './core/result'
 export type * from './core/use-case'
 export * from './core/uuid'
-export * from './core/tax-id/tax-id'
-export * from './core/masks/brazilian-masks'
+export * from './utils/brazilian-masks'
+export * from './utils/tax-id'
 
 export * from './controllers/base-controller'
 export type * from './controllers/controller-errors'
